move pages to own folder
