+++ MauiMoneyMate/AppShell.xaml
@@ -4,11 +4,12 @@
     xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
     xmlns:local="clr-namespace:MauiMoneyMate"
+    xmlns:pages="using:MauiMoneyMate.Pages"
     Shell.FlyoutBehavior="Disabled">
 
     <ShellContent
         Title="Financial Overview"
-        ContentTemplate="{DataTemplate local:MainPage}"
+        ContentTemplate="{DataTemplate pages:MainPage}"
         Route="MainPage"/>
 
 </Shell>
