+++ MauiMoneyMate/MauiProgram.cs
@@ -1,4 +1,5 @@
 ﻿using CommunityToolkit.Maui;
+using MauiMoneyMate.Pages;
 using MauiMoneyMate.Popups;
 using MauiMoneyMate.ViewModels;
