+++ MauiMoneyMate/Pages/FilePage.xaml
@@ -2,7 +2,7 @@
 <ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
              xmlns:viewModels="clr-namespace:MauiMoneyMate.ViewModels"
-             x:Class="MauiMoneyMate.FilePage"
+             x:Class="MauiMoneyMate.Pages.FilePage"
              x:DataType="viewModels:FileViewModel"
              Title="{Binding FilePageTitle}"
              BackgroundColor="{AppThemeBinding Light={StaticResource TertiaryLight}, Dark={StaticResource TertiaryDark}}">
