+++ MauiMoneyMate/Platforms/Windows/App.xaml
@@ -1,5 +1,5 @@
 ﻿<maui:MauiWinUIApplication
-    x:Class="MauiMoneyMate.WinUI.App"
+    x:Class="MauiMoneyMate.Platforms.Windows.App"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:maui="using:Microsoft.Maui"
