+++ MauiMoneyMate/Platforms/Windows/App.xaml.cs
@@ -1,9 +1,9 @@
-﻿using Microsoft.UI.Xaml;
+﻿
 
 // To learn more about WinUI, the WinUI project structure,
 // and more about our project templates, see: http://aka.ms/winui-project-info.
 
-namespace MauiMoneyMate.WinUI
+namespace MauiMoneyMate.Platforms.Windows
 {
     /// <summary>
     /// Provides application-specific behavior to supplement the default Application class.
