+++ MauiMoneyMate/ViewModels/MainViewModel.cs
@@ -5,6 +5,7 @@ using CommunityToolkit.Maui.Alerts;
 using CommunityToolkit.Maui.Views;
 using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
+using MauiMoneyMate.Pages;
 using MauiMoneyMate.Resources.Languages;
 using MauiMoneyMate.Popups;
 using MauiMoneyMate.Utils;
